Move common.statuses to common.http.statuses
