+++ starlette_web/common/utils.py
@@ -9,7 +9,7 @@ from starlette.responses import JSONResponse
 from webargs_starlette import WebargsHTTPException
 
 from starlette_web.core import settings
-from starlette_web.common.statuses import ResponseStatus
+from starlette_web.common.http.statuses import ResponseStatus
 from starlette_web.common.http.exceptions import BaseApplicationError, SendRequestError
