+++ starlette_web/tests/api/test_auth.py
@@ -6,7 +6,7 @@ from datetime import datetime, timedelta
 import pytest
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from starlette_web.common.statuses import ResponseStatus
+from starlette_web.common.http.statuses import ResponseStatus
 from starlette_web.core import settings
 from starlette_web.auth.models import User, UserSession, UserInvite
 from starlette_web.auth.utils import (
